Partial upgrade to 3.38.0
**Changelog:**

- Add overrides for Gtk.Button set/get_focus_on_click
- Python 2 is no longer supported
- gdk overrides: Fix wrapping of scroll events
- meson: Make the `pycairo` option a feature

**Packager Notes:**

We are dual-sourcing this package for 3.38.0 and 3.36.1, as 3.38.0 does not support Python 2, which is required by many packages including our own Software Center.

To resolve this, we are installing 3.36.1 as the python2 lib and 3.38.0 as the python3 lib. Excluding our software center, nothing should be using python2 at this point, so don't worry about building for it.